Moving options/settings into data project
