+++ src/Randomizer.App/ColorBrushConverter.cs
@@ -1,17 +1,24 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Globalization;
 using System.Windows.Data;
 using System.Windows.Media;
 
 namespace Randomizer.App
 {
-    [ValueConversion(typeof(Color), typeof(SolidColorBrush))]
+    [ValueConversion(typeof(List<byte>), typeof(SolidColorBrush))]
     internal class ColorBrushConverter : IValueConverter
     {
         public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
-            => new SolidColorBrush((Color)value);
+        {
+            var rgb = value as List<byte>;
+            return new SolidColorBrush(Color.FromArgb(rgb[0], rgb[1], rgb[2], rgb[3]));
+        }
 
         public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
-            => ((SolidColorBrush)value).Color;
+        {
+            var color = (value as SolidColorBrush).Color;
+            return new List<byte>() { color.A, color.R, color.G, color.B };
+        }
     }
 }
