+++ src/Randomizer.App/RomGenerator.cs
@@ -15,6 +15,7 @@ using Microsoft.Extensions.Logging;
 using Randomizer.App.Patches;
 using Randomizer.App.ViewModels;
 using Randomizer.Data;
+using Randomizer.Data.Options;
 using Randomizer.Data.WorldData.Regions;
 using Randomizer.Shared;
 using Randomizer.Shared.Models;
