+++ src/Randomizer.App/Windows/OptionsWindow.xaml.cs
@@ -13,6 +13,7 @@ using Microsoft.Extensions.Logging;
 using Randomizer.App.ViewModels;
 using Randomizer.SMZ3.ChatIntegration;
 using Randomizer.Data.Configuration;
+using Randomizer.Data.Options;
 
 namespace Randomizer.App
 {
