+++ src/Randomizer.App/Windows/TrackerWindow.xaml.cs
@@ -32,6 +32,7 @@ using Randomizer.Data.Configuration.ConfigFiles;
 using Randomizer.Data.Configuration.ConfigTypes;
 using Randomizer.SMZ3.Tracking.Services;
 using Randomizer.SMZ3.Tracking.VoiceCommands;
+using Randomizer.Data.Options;
 
 namespace Randomizer.App
 {
@@ -630,7 +631,7 @@ namespace Randomizer.App
 
         private void Window_Loaded(object sender, RoutedEventArgs e)
         {
-            Background = new SolidColorBrush(Options.GeneralOptions.TrackerBackgroundColor);
+            Background = new SolidColorBrush(Color.FromArgb(Options.GeneralOptions.TrackerBGColor[0], Options.GeneralOptions.TrackerBGColor[1], Options.GeneralOptions.TrackerBGColor[2], Options.GeneralOptions.TrackerBGColor[3]));
 
             // If a rom was passed in, generate its seed to populate all locations and items
             if (GeneratedRom.IsValid(Rom))
@@ -777,7 +778,7 @@ namespace Randomizer.App
             };
             _autoTrackerDisableMenuItem.Click += (sender, e) =>
             {
-                Tracker.AutoTracker?.SetConnector(SMZ3.Tracking.AutoTracking.EmulatorConnectorType.None);
+                Tracker.AutoTracker?.SetConnector(EmulatorConnectorType.None);
             };
             menu.Items.Add(_autoTrackerDisableMenuItem);
 
@@ -788,7 +789,7 @@ namespace Randomizer.App
             };
             _autoTrackerLuaMenuItem.Click += (sender, e) =>
             {
-                Tracker.AutoTracker?.SetConnector(SMZ3.Tracking.AutoTracking.EmulatorConnectorType.Lua);
+                Tracker.AutoTracker?.SetConnector(EmulatorConnectorType.Lua);
             };
             menu.Items.Add(_autoTrackerLuaMenuItem);
 
@@ -799,7 +800,7 @@ namespace Randomizer.App
             };
             _autoTrackerUSB2SNESMenuItem.Click += (sender, e) =>
             {
-                Tracker.AutoTracker?.SetConnector(SMZ3.Tracking.AutoTracking.EmulatorConnectorType.USB2SNES);
+                Tracker.AutoTracker?.SetConnector(EmulatorConnectorType.USB2SNES);
             };
             menu.Items.Add(_autoTrackerUSB2SNESMenuItem);
 
@@ -858,9 +859,9 @@ namespace Randomizer.App
             StatusBarAutoTrackerDisabled.Visibility = !Tracker.AutoTracker.IsEnabled ? Visibility.Visible : Visibility.Collapsed;
             StatusBarAutoTrackerEnabled.Visibility = Tracker.AutoTracker.IsEnabled && !Tracker.AutoTracker.IsConnected ? Visibility.Visible : Visibility.Collapsed;
             StatusBarAutoTrackerConnected.Visibility = Tracker.AutoTracker.IsEnabled && Tracker.AutoTracker.IsConnected ? Visibility.Visible : Visibility.Collapsed;
-            _autoTrackerDisableMenuItem.IsChecked = Tracker.AutoTracker?.ConnectorType == SMZ3.Tracking.AutoTracking.EmulatorConnectorType.None;
-            _autoTrackerLuaMenuItem.IsChecked = Tracker.AutoTracker?.ConnectorType == SMZ3.Tracking.AutoTracking.EmulatorConnectorType.Lua;
-            _autoTrackerUSB2SNESMenuItem.IsChecked = Tracker.AutoTracker?.ConnectorType == SMZ3.Tracking.AutoTracking.EmulatorConnectorType.USB2SNES;
+            _autoTrackerDisableMenuItem.IsChecked = Tracker.AutoTracker?.ConnectorType == EmulatorConnectorType.None;
+            _autoTrackerLuaMenuItem.IsChecked = Tracker.AutoTracker?.ConnectorType == EmulatorConnectorType.Lua;
+            _autoTrackerUSB2SNESMenuItem.IsChecked = Tracker.AutoTracker?.ConnectorType == EmulatorConnectorType.USB2SNES;
         }
 
         private void UpdateStats(TrackerEventArgs e)
