+++ src/Randomizer.Data/Rdc.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Text;
-
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
