+++ src/Randomizer.Data/ShipSprite.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Randomizer.App
+namespace Randomizer.Data
 {
     public record ShipSprite(string DisplayName, string FileName)
     {
