+++ src/Randomizer.SMZ3.Tracking/AutoTracking/AutoTracker.cs
@@ -12,6 +12,7 @@ using Randomizer.SMZ3.Tracking.AutoTracking.MetroidStateChecks;
 using Randomizer.SMZ3.Tracking.AutoTracking.ZeldaStateChecks;
 using Randomizer.SMZ3.Tracking.Services;
 using Randomizer.SMZ3.Tracking.VoiceCommands;
+using Randomizer.Data.Options;
 
 namespace Randomizer.SMZ3.Tracking.AutoTracking
 {
