+++ src/Randomizer.SMZ3.Tracking/Tracker.cs
@@ -32,6 +32,8 @@ using Randomizer.Data.WorldData.Regions;
 using Randomizer.Data.WorldData.Regions.Zelda;
 using Randomizer.Data.Logic;
 using Randomizer.Data.WorldData;
+using Randomizer.Data;
+using Randomizer.Data.Options;
 
 namespace Randomizer.SMZ3.Tracking
 {
